backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
pillow==10.1.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
